+++ spec/overcommit/hook_context/pre_commit_spec.rb
@@ -487,6 +487,31 @@ describe Overcommit::HookContext::PreCommit do
 
       it { should =~ [File.expand_path('some-file'), File.expand_path('other-file')] }
     end
+
+    context 'when changing a symlink to a directory during an amendment' do
+      around do |example|
+        repo do
+          `git commit --allow-empty -m "Initial commit"`
+          FileUtils.mkdir 'some-directory'
+          FileUtils.ln_s 'some-directory', 'some-symlink'
+          `git add some-symlink some-directory`
+          `git commit -m "Add file"`
+          `git rm some-symlink`
+          FileUtils.mkdir 'some-symlink'
+          FileUtils.touch File.join('some-symlink', 'another-file')
+          `git add some-symlink`
+          example.run
+        end
+      end
+
+      before do
+        context.stub(:amendment?).and_return(true)
+      end
+
+      it 'does not include the directory in the list of modified files' do
+        subject.should_not include File.expand_path('some-symlink')
+      end
+    end
   end
 
   describe '#modified_lines_in_file' do